Add documentation and cleanup code
