+++ src/checkers/inference/solver/backend/encoder/binary/ComparableConstraintEncoder.java
@@ -1,7 +1,12 @@
 package checkers.inference.solver.backend.encoder.binary;
 
+import checkers.inference.util.ConstraintVerifier;
+
 /**
- * Created by mier on 07/11/17.
+ * A marker interface that all constraint encoder that supports encoding comparable constraint
+ * should implement. Otherwise, your encoder will be considered not supporting encoding comparable
+ * constraint and rejected by the FormatTranslator factory method
+ * @see checkers.inference.solver.backend.AbstractFormatTranslator#createComparableConstraintEncoder(ConstraintVerifier)
  */
 public interface ComparableConstraintEncoder<ConstraintEncodingT> extends BinaryConstraintEncoder<ConstraintEncodingT> {
 }
