+++ src/checkers/inference/solver/backend/encoder/binary/EqualityConstraintEncoder.java
@@ -1,7 +1,12 @@
 package checkers.inference.solver.backend.encoder.binary;
 
+import checkers.inference.util.ConstraintVerifier;
+
 /**
- * Created by mier on 07/11/17.
+ * A marker interface that all constraint encoder that supports encoding equality constraint
+ * should implement. Otherwise, your encoder will be considered not supporting encoding equality
+ * constraint and rejected by the FormatTranslator factory method
+ * @see checkers.inference.solver.backend.AbstractFormatTranslator#createEqualityConstraintEncoder(ConstraintVerifier)
  */
 public interface EqualityConstraintEncoder<ConstraintEncodingT> extends BinaryConstraintEncoder<ConstraintEncodingT> {
 }
