+++ src/checkers/inference/solver/backend/logiql/encoder/LogiQLEqualityConstraintEncoder.java
@@ -7,9 +7,6 @@ import checkers.inference.solver.frontend.Lattice;
 import checkers.inference.solver.util.NameUtils;
 import checkers.inference.util.ConstraintVerifier;
 
-/**
- * Created by mier on 07/11/17.
- */
 public class LogiQLEqualityConstraintEncoder extends LogiQLAbstractConstraintEncoder implements EqualityConstraintEncoder<String> {
 
     public LogiQLEqualityConstraintEncoder(Lattice lattice, ConstraintVerifier verifier) {
