+++ src/checkers/inference/solver/backend/logiql/encoder/LogiQLSubtypeConstraintEncoder.java
@@ -7,9 +7,6 @@ import checkers.inference.solver.frontend.Lattice;
 import checkers.inference.solver.util.NameUtils;
 import checkers.inference.util.ConstraintVerifier;
 
-/**
- * Created by mier on 07/11/17.
- */
 public class LogiQLSubtypeConstraintEncoder extends LogiQLAbstractConstraintEncoder implements SubtypeConstraintEncoder<String> {
 
     public LogiQLSubtypeConstraintEncoder(Lattice lattice, ConstraintVerifier verifier) {
